Fixing matview generation to be after loading initial table data
