+++ usaspending_api/conftest.py
@@ -1,7 +1,7 @@
 from django.conf import settings
-import subprocess
 import pytest
-from django.db import connection
+
+from usaspending_api.common.helpers import generate_matviews
 
 
 def pytest_configure():
@@ -13,24 +13,6 @@ def pytest_configure():
     settings.DATABASES['default'] = test_db
     # Also remove any database routers
     settings.DATABASE_ROUTERS.clear()
-
-
-MATVIEW_GENERATOR_FILE = "usaspending_api/database_scripts/matview_generator/matview_sql_generator.py"
-
-
-ENUM_FILE = ['usaspending_api/database_scripts/matviews/functions_and_enums.sql']
-
-
-TEMP_SQL_FILES = ['../matviews/universal_transaction_matview.sql',
-                  '../matviews/universal_award_matview.sql',
-                  '../matviews/summary_transaction_view.sql',
-                  '../matviews/summary_transaction_month_view.sql',
-                  '../matviews/summary_transaction_geo_view.sql',
-                  '../matviews/summary_award_view.sql',
-                  '../matviews/summary_view_cfda_number.sql',
-                  '../matviews/summary_view_naics_codes.sql',
-                  '../matviews/summary_view_psc_codes.sql',
-                  '../matviews/summary_view.sql']
 
 
 def pytest_addoption(parser):
@@ -58,11 +40,6 @@ def django_db_setup(django_db_blocker,
                 interactive=False,
                 **setup_databases_args
             )
-        with connection.cursor() as c:
-                c.execute(get_sql(ENUM_FILE)[0])
-                subprocess.call("python  " + MATVIEW_GENERATOR_FILE, shell=True)
-                for file in get_sql(TEMP_SQL_FILES):
-                    c.execute(file)
 
     def teardown_database():
             with django_db_blocker.unblock():
@@ -75,9 +52,6 @@ def django_db_setup(django_db_blocker,
         request.addfinalizer(teardown_database)
 
 
-def get_sql(sql_files):
-    data = []
-    for file in sql_files:
-        with open(file, 'r') as myfile:
-            data.append(myfile.read())
-    return data
+@pytest.fixture()
+def refresh_matviews():
+    generate_matviews()
